+++ frontend/src/AgentGames/Feedback/GameResultsWrapper.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
-import ResultsDisplay from '../Utilities/ResultsDisplay';
+import ResultsDisplay from "../Shared/Utilities/ResultsDisplay";
 import Lineup4ResultsDisplay from './games/lineup4/Lineup4ResultsDisplay';
 
 const GameResultsWrapper = (props) => {
