+++ frontend/src/AgentGames/Shared/Utilities/InstructionPopup.jsx
