+++ frontend/src/AgentGames/Shared/Utilities/ResultsDisplay.jsx
