+++ frontend/src/AgentGames/Shared/Utilities/UserTooltips.jsx
