+++ frontend/src/App.jsx
@@ -1,7 +1,7 @@
 import AgentGamesNavbar from './Navbar';
 import AgentHome from './AgentGames/AgentHome';
 import AgentLogin from './AgentGames/User/AgentLogin';
-import AgentRankings from './AgentGames/Utilities/Rankings';
+import AgentRankings from "./AgentGames/Shared/Utilities/Rankings";
 import AgentSubmission from './AgentGames/User/AgentSubmission';
 import AgentLeagueSignUp from './AgentGames/User/LeagueSignup';
 import Demo from './AgentGames/Demo';
